block edit/delete when a simulation is in progress

+++ WebUI/src/components/SimulationsTable/SimulationsTable.js
@@ -4,6 +4,7 @@ import {Cell} from '@enact/ui/Layout';
 import css from './SimulationsTable.module.less';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
+const blockingAction = (status) => ['Running', 'Initializing', 'Deinitializing'].includes(status);
 
 class SimulationsTable extends React.Component {
     constructor(props) {
@@ -24,24 +25,27 @@ class SimulationsTable extends React.Component {
 
     openEdit = (ev) => {
         const id = parseInt(ev.currentTarget.dataset.simulationid);
-        this.props.openEdit(id);
+        const status = this.props.simulations.get(id).status;
+        if (!blockingAction(status)) this.props.openEdit(id);
     }
 
     handleDelete = (ev) => {
         const id = parseInt(ev.currentTarget.dataset.simulationid);
-        this.props.handleDelete(id);
+        const status = this.props.simulations.get(id).status;
+        if (!blockingAction(status)) this.props.handleDelete(id);
     }
 
     simulationList() {
         const list = [];
         for (const [i, simulation] of this.props.simulations) {
             const classes = classNames(css.simulationItem, {[css.selected]: this.props.selected === i});
+            const btnClassNames = classNames({[css.disabled]: blockingAction(simulation.status)})
             list.push(
                 <tr key={`${simulation}-${i}`} className={classes} data-simulationid={i}>
                     <td data-simulationid={i} onClick={this.selectSimulation}>{simulation.name}</td>
                     <td>{simulation.status}</td>
-                    <td data-simulationid={simulation.id} onClick={this.openEdit}><FaRegEdit /></td>
-                    <td data-simulationid={simulation.id} onClick={this.handleDelete}><FaRegWindowClose /></td>
+                    <td data-simulationid={simulation.id} onClick={this.openEdit}><FaRegEdit className={btnClassNames} /></td>
+                    <td data-simulationid={simulation.id} onClick={this.handleDelete}><FaRegWindowClose className={btnClassNames} /></td>
                 </tr>
             )
         }
